+++ modules/juce_graphics/geometry/juce_EdgeTable.h
@@ -129,7 +129,7 @@ public:
                 while (--numPoints >= 0)
                 {
                     const int level = *++line;
-                    jassert (isPositiveAndBelow (level, (int) 256));
+                    jassert (isPositiveAndBelow (level, 256));
                     const int endX = *++line;
                     jassert (endX >= x);
                     const int endOfRun = (endX >> 8);
